feat(demo): comment node support more button
