+++ apps/demo-free-layout/src/components/comment/components/index.css
@@ -68,3 +68,37 @@
     resize: none;
     outline: none;
 }
+
+.workflow-comment-more-button {
+    position: absolute;
+    right: 6px;
+}
+
+.workflow-comment-more-button > .semi-button {
+    color: rgba(255, 255, 255, 0);
+    background: none;
+}
+
+.workflow-comment-more-button > .semi-button:hover {
+    color: #ffa100;
+    background: #fbf2d2cc;
+    backdrop-filter: blur(1px);
+}
+
+.workflow-comment-more-button-focused > .semi-button:hover {
+    color: #ff811a;
+    background: #ffe3cecc;
+    backdrop-filter: blur(1px);
+}
+
+.workflow-comment-more-button > .semi-button:active {
+    color: #f2b600;
+    background: #ede5c7cc;
+    backdrop-filter: blur(1px);
+}
+
+.workflow-comment-more-button-focused > .semi-button:active {
+    color: #ff811a;
+    background: #eed5c1cc;
+    backdrop-filter: blur(1px);
+}
