+++ apps/demo-free-layout/src/form-components/form-header/index.tsx
@@ -1,122 +1,17 @@
-import { useCallback, useState, type MouseEvent } from 'react';
-
-import {
-  delay,
-  Field,
-  FieldRenderProps,
-  useClientContext,
-  useService,
-  WorkflowDragService,
-  WorkflowSelectService,
-} from '@flowgram.ai/free-layout-editor';
-import { NodeIntoContainerService } from '@flowgram.ai/free-container-plugin';
-import { IconButton, Dropdown, Typography, Button } from '@douyinfe/semi-ui';
-import { IconMore, IconSmallTriangleDown, IconSmallTriangleLeft } from '@douyinfe/semi-icons';
+import { Field, FieldRenderProps } from '@flowgram.ai/free-layout-editor';
+import { Typography, Button } from '@douyinfe/semi-ui';
+import { IconSmallTriangleDown, IconSmallTriangleLeft } from '@douyinfe/semi-icons';
 
 import { Feedback } from '../feedback';
-import { FlowNodeRegistry } from '../../typings';
-import { PasteShortcut } from '../../shortcuts/paste';
-import { CopyShortcut } from '../../shortcuts/copy';
 import { useIsSidebar, useNodeRenderContext } from '../../hooks';
+import { NodeMenu } from '../../components/node-menu';
 import { getIcon } from './utils';
 import { Header, Operators, Title } from './styles';
 
 const { Text } = Typography;
 
-function DropdownContent() {
-  const [visible, setVisible] = useState(true);
-  const { node, deleteNode } = useNodeRenderContext();
-  const clientContext = useClientContext();
-  const registry = node.getNodeRegistry<FlowNodeRegistry>();
-  const nodeIntoContainerService = useService(NodeIntoContainerService);
-  const selectService = useService(WorkflowSelectService);
-  const dragService = useService(WorkflowDragService);
-  const canMoveOut = nodeIntoContainerService.canMoveOutContainer(node);
-
-  const rerenderMenu = useCallback(() => {
-    // force destroy component - 强制销毁组件触发重新渲染
-    setVisible(false);
-    requestAnimationFrame(() => {
-      setVisible(true);
-    });
-  }, []);
-
-  const handleMoveOut = useCallback(
-    async (e: MouseEvent) => {
-      e.stopPropagation();
-      const sourceParent = node.parent;
-      // move out of container - 移出容器
-      nodeIntoContainerService.moveOutContainer({ node });
-      // clear invalid lines - 清除非法线条
-      await nodeIntoContainerService.clearInvalidLines({
-        dragNode: node,
-        sourceParent,
-      });
-      rerenderMenu();
-      await delay(16);
-      // select node - 选中节点
-      selectService.selectNode(node);
-      // start drag node - 开始拖拽
-      dragService.startDragSelectedNodes(e);
-    },
-    [nodeIntoContainerService, node, rerenderMenu]
-  );
-
-  const handleCopy = useCallback(
-    (e: React.MouseEvent) => {
-      const copyShortcut = new CopyShortcut(clientContext);
-      const pasteShortcut = new PasteShortcut(clientContext);
-      const data = copyShortcut.toClipboardData([node]);
-      pasteShortcut.apply(data);
-      e.stopPropagation(); // Disable clicking prevents the sidebar from opening
-    },
-    [clientContext, node]
-  );
-
-  const handleDelete = useCallback(
-    (e: React.MouseEvent) => {
-      deleteNode();
-      e.stopPropagation(); // Disable clicking prevents the sidebar from opening
-    },
-    [clientContext, node]
-  );
-
-  if (!visible) {
-    return;
-  }
-
-  return (
-    <Dropdown
-      trigger="hover"
-      position="bottomRight"
-      render={
-        <Dropdown.Menu>
-          {canMoveOut && <Dropdown.Item onClick={handleMoveOut}>Move out</Dropdown.Item>}
-          <Dropdown.Item onClick={handleCopy} disabled={registry.meta!.copyDisable === true}>
-            Create Copy
-          </Dropdown.Item>
-          <Dropdown.Item
-            onClick={handleDelete}
-            disabled={!!(registry.canDelete?.(clientContext, node) || registry.meta!.deleteDisable)}
-          >
-            Delete
-          </Dropdown.Item>
-        </Dropdown.Menu>
-      }
-    >
-      <IconButton
-        color="secondary"
-        size="small"
-        theme="borderless"
-        icon={<IconMore />}
-        onClick={(e) => e.stopPropagation()}
-      />
-    </Dropdown>
-  );
-}
-
 export function FormHeader() {
-  const { node, expanded, toggleExpand, readonly } = useNodeRenderContext();
+  const { node, expanded, toggleExpand, readonly, deleteNode } = useNodeRenderContext();
   const isSidebar = useIsSidebar();
   const handleExpand = (e: React.MouseEvent) => {
     toggleExpand();
@@ -147,7 +42,7 @@ export function FormHeader() {
       )}
       {readonly ? undefined : (
         <Operators>
-          <DropdownContent />
+          <NodeMenu node={node} deleteNode={deleteNode} />
         </Operators>
       )}
     </Header>
